+++ guns-admin/src/main/java/com/stylefeng/guns/modular/system/service/IMenuService.java
@@ -14,7 +14,7 @@ public interface IMenuService {
      * @author stylefeng
      * @Date 2017/5/5 22:20
      */
-    void delMenu(Integer menuId);
+    void delMenu(Long menuId);
 
     /**
      * 删除菜单包含所有子菜单
@@ -22,5 +22,5 @@ public interface IMenuService {
      * @author stylefeng
      * @Date 2017/6/13 22:02
      */
-    void delMenuContainSubMenus(Integer menuId);
+    void delMenuContainSubMenus(Long menuId);
 }
